requirements.txt
flask
flask-wtf
flask-login
flask-bcrypt
flask-sqlalchemy
python-dotenv
bcrypt
gunicorn
boto3
email_validator